V8.0.1924 (#113)
* Update ServiceInfoCtrl.java

* refactor(logical rename):

4 minutes

* feat(new messenger-lib):

1 minute

* fix(props table upd):

3 minutes

3 minutes

2 minutes

5 minutes

* fix(Bean creating except. no db connection):

4 minutes

* feat(ftp try active mode):

7 minutes

52 seconds

* fix(ThreadConfig):
feat(PCUserResolver.java): refactor

* fix(file time pref):

* fix(TODO fixes):

* test:

* fix(active|passive ftp):

2 minutes

* test:

18 minutes

* chore:

1 minute

* fix(todo):

13 minutes

* Delete inet.uniq

* test:

* fix(visitor):

* test(AnketaKonfeta):

* test(CalculateCTRL):

* feat:

* chore:

* chore:

* fix(inet stats):

36 minutes

* feat(APP_Build):

2 minutes

* fix:

11 minutes

* fix(current provider):

2 minutes

* fix(VersionInfo):

25 minutes

+++ modules/networker/src/main/java/ru/vachok/networker/statistics/InteretStats.java
@@ -7,7 +7,6 @@ import com.mysql.jdbc.jdbc2.optional.MysqlDataSource;
 import ru.vachok.messenger.MessageSwing;
 import ru.vachok.messenger.MessageToUser;
 import ru.vachok.mysqlandprops.RegRuMysql;
-import ru.vachok.networker.AppComponents;
 import ru.vachok.networker.ConstantsFor;
 import ru.vachok.networker.TForms;
 import ru.vachok.networker.abstr.DataBaseRegSQL;
@@ -18,8 +17,12 @@ import ru.vachok.networker.services.MessageLocal;
 import java.awt.*;
 import java.io.*;
 import java.sql.*;
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
 import java.time.LocalTime;
+import java.util.Date;
 import java.util.List;
+import java.util.concurrent.Executors;
 
 
 /**
@@ -53,7 +56,7 @@ public class InteretStats implements Runnable, DataBaseRegSQL {
         long iPsWithInet = readIPsWithInet();
         messageToUser.info(getClass().getSimpleName() + "in kbytes. ", new File(FILENAME_INETSTATSIPCSV).getAbsolutePath(), " = " + iPsWithInet);
         readInetStatsRSetToCSV();
-        AppComponents.threadConfig().execByThreadConfig(new InetStatSorter());
+        Executors.unconfigurableExecutorService(Executors.newSingleThreadExecutor()).execute(new InetStatSorter());
     }
     
     @Override public int selectFrom() {
@@ -148,6 +151,8 @@ public class InteretStats implements Runnable, DataBaseRegSQL {
     
     private void readInetStatsRSetToCSV() {
         List<String> chkIps = FileSystemWorker.readFileToList(new File(FILENAME_INETSTATSIPCSV).getPath());
+        DateFormat format = new SimpleDateFormat("E");
+        String weekDay = format.format(new Date());
         long totalBytes = 0;
         for (String ip : chkIps) {
             this.fileName = FILENAME_INETSTATSCSV.replace("inetstats", ip).replace(".csv", "_" + LocalTime.now().toSecondOfDay() + ".csv");
@@ -157,7 +162,7 @@ public class InteretStats implements Runnable, DataBaseRegSQL {
             totalBytes += file.length();
             new MessageLocal(getClass().getSimpleName())
                 .info(fileName, file.length() / ConstantsFor.KBYTE + " kb", "total kb: " + totalBytes / ConstantsFor.KBYTE);
-            if (file.length() > 10) {
+            if (weekDay.equals("вс") && file.length() > 10) {
                 this.sql = new StringBuilder().append("DELETE FROM `inetstats` WHERE `ip` LIKE '").append(ip).append("'").toString();
                 System.out.println(deleteFrom() + " rows deleted.");
             }
